requirements.txt
asgiref==3.3.1
attrs==20.3.0
Automat==20.2.0
certifi==2020.11.8
chardet==3.0.4
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
Django==3.1.3
django-filter==2.4.0
django-guardian==2.3.0
djangorestframework==3.12.2
hyperlink==20.0.1
idna==2.10
incremental==17.5.0
itypes==1.2.0
Jinja2==2.11.2
Markdown==3.3.3
MarkupSafe==1.1.1
Pillow==8.0.1
Pygments==2.7.2
PyHamcrest==2.0.2
pytz==2020.4
requests==2.25.0
six==1.15.0
sqlparse==0.4.1
Twisted==20.3.0
uritemplate==3.0.1
urllib3==1.26.2
zope.interface==5.2.0
